DSM-0064   tutorial level messages
NOTE - the MessageSpot.java was a failed experiment that made it into one of the builds.
There is a slight chance that saved Tutorial level games from the interim build may die
horribly when reloaded... so I am going to wait until after I either test the removal fully
or have given enough time so that (virtually) no one will be effected prior to removal.
You should not pick up this file in your builds unless you want to play with the functionality.